Create character (#14)
* 不要なので削除

* パスの変更

* enumから数値で変動するよう変更

* 定数名の変更

* 不要なデータの削除

* 修正

* ボスにステートを追加

* インクルードを修正

* BasicEnemyの追加

* DeformEnemyの追加

* 不要な処理を削除

* サウンドマネージャーを修正

* エネミークラスの追加

* 文字コードの変更

* UITextを実装

* ギアの獲得数表示の関数を編集

* オブジェクトの追加

* SetTransformの追加

* オブジェクトの追加

* スポナーの修正

* コメント修正

* プレイヤーのステートマシンを一部修正

* エネミーのステートマシンを修正

* 一旦commit

* ステートマシンの変数をprivateに

* 一旦コミット

* リザルトシーケンス中はUIを非表示にする

* intとuint8_tの書き換え

* バトルマネージャーの勝敗enumをclassに変更

* ロード画面のステートenumをclassに変更

* コリジョンタイプのenumをクラスにした

* ステートの名前を変更

* ファイル名変更

* プレイヤーのアニメーション機構を変更

* エネミーのアニメーション再生処理の更新

* アニメーションクリップのenumをactertypesに移し、classに変更

* 不要なインクルードを削除

* ステートマシン関連の変更

* ステートマシンを分割

* ステートマシンの修正

* actorstatemachineの追加

* ステートマシンの移動

* 各キャラのステートマシンを実装

* エネミーのステートマシンを分割

* エネミーのステートマシンを分割

* インクルードの変更

* 勝敗判定の変更

* インクルードの変更

* ダメージ処理の変更

* ダメージ処理の追加

* アクターとキャラクターにそれぞれステータスとステートマシンの変数を用意

+++ Game/Core/ParameterManager.h
@@ -27,7 +27,7 @@ struct MasterPlayerStatusParameter : public IMasterParameter
 {
 	appParameter(MasterPlayerStatusParameter);
 	//
-	uint8_t maxHp;
+	int maxHp;
 	float walkSpeed;
 	float dashSpeed;
 	float hurtRadius;
@@ -39,7 +39,7 @@ struct MasterBasicEnemyStatusParameter : public IMasterParameter
 {
 	appParameter(MasterBasicEnemyStatusParameter);
 	//
-	uint8_t maxHp;
+	int maxHp;
 	float walkSpeed;
 	float dashSpeed;
 	float hurtRadius;
@@ -51,7 +51,7 @@ struct MasterDeformEnemyStatusParameter : public IMasterParameter
 {
 	appParameter(MasterDeformEnemyStatusParameter);
 	//
-	uint8_t maxHp;
+	int maxHp;
 	float walkSpeed;
 	float dashSpeed;
 	float hurtRadius;
@@ -64,7 +64,7 @@ struct MasterBossEnemyStatusParameter : public IMasterParameter
 {
 	appParameter(MasterBossEnemyStatusParameter);
 	//
-	uint8_t maxHp;
+	int maxHp;
 	float walkSpeed;
 	float dashSpeed;
 	float hurtRadius;
@@ -144,7 +144,7 @@ public:
 	 * データを取得する関数
 	 */
 	template <typename T>
-	const T* GetParameter(int index = 0) const
+	const T* GetParameter(uint8_t index = 0) const
 	{
 		auto it = m_parameterMap.find(T::ID());
 		if (it != m_parameterMap.end()) {
